# openapi.yaml
openapi: 3.0.0
info:
  title: QuantifiedSelf API
  version: '1.2'
  description: This api is made by Abhay Chaudhary to access QuantifiedSelf App with the help of API access.
  contact:
    name: Abhay Chaudhary
servers:
  - url: 'http://127.0.0.1:5000'
    description: localhost
paths:
  /api/login:
    parameters: []
    post:
      summary: ''
      operationId: post-login
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                description: ''
                type: object
                x-examples:
                  example-1:
                    access_token: hsjjgdfwydgf
                properties:
                  access_token:
                    type: string
                    minLength: 1
                  name:
                    type: string
                required:
                  - access_token
                  - name
              examples:
                example-1:
                  value:
                    name: user1
                    access_token: string
        '400':
          description: Bad Request; including validation errors
          content:
            application/json:
              schema:
                description: ''
                type: object
                properties:
                  msg:
                    type: string
                    minLength: 1
                required:
                  - msg
                x-examples:
                  example-1:
                    msg: success
              examples:
                example-1:
                  value:
                    msg: error occurred
        '401':
          description: Invalid username or password supplied
          content:
            application/json:
              schema:
                description: ''
                type: object
                properties:
                  msg:
                    type: string
                    minLength: 1
                required:
                  - msg
                x-examples:
                  example-1:
                    msg: Bad username or password
              examples:
                login_error_401:
                  value:
                    msg: Bad username or password
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                description: ''
                type: object
                properties:
                  msg:
                    type: string
                    minLength: 1
                required:
                  - msg
                x-examples:
                  example-1:
                    msg: Internal server error occurred
              examples:
                Internal Server Error:
                  value:
                    msg: Internal server error occurred
      description: Login user and get JWT Authentication Token
      requestBody:
        content:
          application/json:
            schema:
              description: ''
              type: object
              x-examples:
                example-1:
                  username: test@test.com
                  password: mypass
              properties:
                email:
                  type: string
                  minLength: 1
                password:
                  type: string
                  minLength: 1
              required:
                - email
                - password
            examples:
              example-1:
                value:
                  email: test@test.com
                  password: mypassword
...
  /api/register:
    post:
      summary: ''
      operationId: post-api-register
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                description: ''
                type: object
                properties:
                  msg:
                    type: string
                    minLength: 1
                required:
                  - msg
                x-examples:
                  example-1:
                    msg: created
              examples:
                example-1:
                  value:
                    msg: created
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                description: ''
                type: object
                properties:
                  msg:
                    type: string
                    minLength: 1
                required:
                  - msg
                x-examples:
                  example-1:
                    msg: created
              examples:
                email_exists:
                  value:
                    msg: email already exists
        '500':
          description: Internal Server Error
      description: create a new user in the database
      requestBody:
        content:
          application/json:
            schema:
              description: ''
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                email:
                  type: string
                  minLength: 1
                password:
                  type: string
                  minLength: 1
              required:
                - name
                - email
                - password
              x-examples:
                example-1:
                  name: Abhay Chaudhary
                  email: abhay@test.com
                  password: abhaychaudhary
            examples:
              example-1:
                value:
                  name: Abhay Chaudhary
                  email: abhay1@test.com
                  password: abhaychaudhary
...
components:
...
  securitySchemes:
    api_key_jwt:
      type: http
      scheme: bearer
security:
  - api_key_jwt: []